+++ crates/miden-objects/src/errors.rs
@@ -281,16 +281,51 @@ pub enum AccountTreeError {
 
 #[derive(Debug, Error)]
 pub enum AddressError {
-    #[error("tag length {0} should be {expected} bits for network accounts", expected = crate::note::NoteTag::DEFAULT_NETWORK_TAG_LENGTH)]
+    #[error("tag length {0} should be {expected} bits for network accounts",
+        expected = NoteTag::DEFAULT_NETWORK_TAG_LENGTH
+    )]
     CustomTagLengthNotAllowedForNetworkAccounts(u8),
-    #[error("tag length {0} is too large, must be less than or equal to {max}", max = crate::note::NoteTag::MAX_LOCAL_TAG_LENGTH)]
+    #[error("tag length {0} is too large, must be less than or equal to {max}",
+        max = NoteTag::MAX_LOCAL_TAG_LENGTH
+    )]
     TagLengthTooLarge(u8),
     #[error("unknown address interface `{0}`")]
     UnknownAddressInterface(u16),
     #[error("failed to decode account ID")]
     AccountIdDecodeError(#[source] AccountIdError),
+    #[error("address separator must not be included without routing parameters")]
+    TrailingSeparator,
     #[error("failed to decode bech32 string into an address")]
     Bech32DecodeError(#[source] Bech32Error),
+    #[error("{error_msg}")]
+    DecodeError {
+        error_msg: Box<str>,
+        // thiserror will return this when calling Error::source on NoteError.
+        source: Option<Box<dyn Error + Send + Sync + 'static>>,
+    },
+    #[error("found unknown routing parameter key {0}")]
+    UnknownRoutingParameterKey(u8),
+}
+
+impl AddressError {
+    /// Creates an [`AddressError::DecodeError`] variant from an error message.
+    pub fn decode_error(message: impl Into<String>) -> Self {
+        let message: String = message.into();
+        Self::DecodeError { error_msg: message.into(), source: None }
+    }
+
+    /// Creates an [`AddressError::DecodeError`] variant from an error message and
+    /// a source error.
+    pub fn decode_error_with_source(
+        message: impl Into<String>,
+        source: impl Error + Send + Sync + 'static,
+    ) -> Self {
+        let message: String = message.into();
+        Self::DecodeError {
+            error_msg: message.into(),
+            source: Some(Box::new(source)),
+        }
+    }
 }
 
 // BECH32 ERROR